Split BaseLayout into Nav and Sidebar
BaseLayout was complex in terms of both understanding and developing.
So, split it into simpler smaller components
