+++ dashboard/v1.1/src/layouts/BaseLayout.tsx
@@ -1,29 +1,9 @@
 import { CssBaseline } from '@material-ui/core';
-import {
-  AppBar,
-  Badge,
-  Container,
-  Divider,
-  Drawer,
-  IconButton,
-  List,
-  ListItem,
-  ListItemIcon,
-  ListItemText,
-  Toolbar,
-  Typography
-} from '@material-ui/core';
+import { Container } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
-import {
-  ChevronLeft as ChevronLeftIcon,
-  Dashboard as DashboardIcon,
-  Menu as MenuIcon,
-  Notifications as NotificationsIcon
-} from '@material-ui/icons';
-import clsx from 'clsx';
 import React, { ReactElement, useState } from 'react';
-
-const drawerWidth = 240;
+import Navbar from './Navbar';
+import Sidebar from './Sidebar';
 
 const useStyles = makeStyles(theme => ({
   root: {
@@ -31,64 +11,6 @@ const useStyles = makeStyles(theme => ({
   },
   color: {
     primary: theme.palette.primary
-  },
-  // AppBar styles
-  appBar: {
-    zIndex: theme.zIndex.drawer + 1,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen
-    })
-  },
-  appBarShift: {
-    marginLeft: drawerWidth,
-    width: `calc(100% - ${drawerWidth}px)`,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen
-    })
-  },
-  // Toolbar styles
-  toolbar: {
-    paddingRight: 24 // keep right padding when drawer closed
-  },
-  toolbarIcon: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'flex-end',
-    padding: '0 8px',
-    ...theme.mixins.toolbar
-  },
-  // IconMenu styles
-  menuButton: {
-    marginRight: 36
-  },
-  menuButtonHidden: {
-    display: 'none'
-  },
-  title: {
-    flexGrow: 1
-  },
-  // Drawer styles
-  drawerPaper: {
-    position: 'relative',
-    whiteSpace: 'nowrap',
-    width: drawerWidth,
-    transition: theme.transitions.create('width', {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen
-    })
-  },
-  drawerPaperClose: {
-    overflowX: 'hidden',
-    transition: theme.transitions.create('width', {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen
-    }),
-    width: theme.spacing(7),
-    [theme.breakpoints.up('sm')]: {
-      width: theme.spacing(9)
-    }
   },
   // Content styles
   appBarSpacer: theme.mixins.toolbar,
@@ -115,86 +37,12 @@ export default function BaseLayout(props: any): ReactElement {
   const handleDrawerClose = () => {
     setOpen(false);
   };
-  // Navbar element
-  const nav = (
-    <AppBar
-      position="absolute"
-      className={clsx(classes.appBar, open && classes.appBarShift)}
-    >
-      <Toolbar className={classes.toolbar}>
-        <IconButton
-          edge="start"
-          color="inherit"
-          aria-label="open drawer"
-          onClick={handleDrawerOpen}
-          className={clsx(classes.menuButton, open && classes.menuButtonHidden)}
-        >
-          <MenuIcon />
-        </IconButton>
-        <Typography
-          component="h1"
-          variant="h6"
-          color="inherit"
-          noWrap={true}
-          className={classes.title}
-        >
-          Bench Routes
-        </Typography>
-        <IconButton color="inherit">
-          <Badge badgeContent={4} color="secondary">
-            <NotificationsIcon />
-          </Badge>
-        </IconButton>
-      </Toolbar>
-    </AppBar>
-  );
-
-  // Sidebar element
-  const menuItems = (
-    <div>
-      <ListItem button={true}>
-        <ListItemIcon>
-          <DashboardIcon />
-        </ListItemIcon>
-        <ListItemText primary="Dashboard" />
-      </ListItem>
-      <ListItem button={true}>
-        <ListItemIcon>
-          <DashboardIcon />
-        </ListItemIcon>
-        <ListItemText primary="Dashboard" />
-      </ListItem>
-      <ListItem button={true}>
-        <ListItemIcon>
-          <DashboardIcon />
-        </ListItemIcon>
-        <ListItemText primary="Dashboard" />
-      </ListItem>
-    </div>
-  );
-  const sideBar = (
-    <Drawer
-      variant="permanent"
-      classes={{
-        paper: clsx(classes.drawerPaper, !open && classes.drawerPaperClose)
-      }}
-      open={open}
-    >
-      <div className={classes.toolbarIcon}>
-        <IconButton onClick={handleDrawerClose}>
-          <ChevronLeftIcon />
-        </IconButton>
-      </div>
-      <Divider />
-      <List>{menuItems}</List>
-    </Drawer>
-  );
 
   return (
     <div className={classes.root}>
       <CssBaseline />
-      {nav}
-      {sideBar}
+      <Navbar handleDrawerOpen={handleDrawerOpen} open={open} />
+      <Sidebar handleDrawerClose={handleDrawerClose} open={open} />
       <main className={classes.content}>
         <div className={classes.appBarSpacer} />
         <Container maxWidth="lg" className={classes.container}>
